// src/type/enum.ts
export enum AppStatus {
    NOT_LOADED = 'NOT_LOADED',
    LOADING = 'LOADING',
    LOADED = 'LOADED',
    BOOTSTRAPPING = 'BOOTSTRAPPING',
    NOT_MOUNTED = 'NOT_MOUNTED',
    MOUNTING = 'MOUNTING',
    MOUNTED = 'MOUNTED',
    UNMOUNTING = 'UNMOUNTING',
}

export enum SandBoxType {
    Proxy = 'Proxy',
    Snapshot = 'Snapshot',
  }